requirements.txt
requests
openpyxl
beautifulsoup4
xls2xlsx
aiogram
google-api-python-client